+++ netket/experimental/driver/tdvp.py
@@ -23,7 +23,12 @@ from netket.driver.vmc_common import info
 from netket.operator import AbstractOperator
 from netket.optimizer import LinearOperator
 from netket.optimizer.qgt import QGTAuto
-from netket.vqs import VariationalState, VariationalMixedState, MCState, ExactState
+from netket.vqs import (
+    VariationalState,
+    VariationalMixedState,
+    MCState,
+    FullSummationState,
+)
 
 from netket.experimental.dynamics import RKIntegratorConfig
 
@@ -135,7 +140,7 @@ class TDVP(TDVPBaseDriver):
 
 @odefun.dispatch
 def odefun_tdvp(  # noqa: F811
-    state: Union[MCState, ExactState], driver: TDVP, t, w, *, stage=0
+    state: Union[MCState, FullSummationState], driver: TDVP, t, w, *, stage=0
 ):
     # pylint: disable=protected-access