+++ netket/hilbert/random/particle.py
@@ -32,7 +32,7 @@ def take_sub(key, x, n):
 @dispatch
 def random_state(hilb: Particle, key, batches: int, *, dtype):
     """If no periodic boundary conditions are present particles are positioned normally distributed around the origin.
-    
+
     If periodic boundary conditions are present the particles are positioned uniformly inside the box and a small
     gaussian noise is added on top.
     If periodic boundary conditions are chosen only for certain dimensions, the periodic initialization is used for